RAG/requirements.txt
fastapi
uvicorn
requests
langchain
openai
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
python-dotenv
PyMuPDF